requirements.txt
nextcord==2.2.0
python-dotenv==0.21.2
googlesearch-python==1.1.0
beautifulsoup4==4.9.3
google-api-python-client==2.52.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
dateparser==1.1.1
pyluach==1.4.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.2
more-itertools==8.14.0
markdown==3.4.1
markdownify==0.11.6
lxml==4.9.1
asyncpg==0.26.0